Backend cleanup done
