--- bookstore-api/app/controllers/books_controller.rb
@@ -1,52 +1,0 @@
-class BooksController < ApplicationController
-  before_action :set_book, only: [:show, :update, :destroy]
-
-  # GET /books
-  def index
-    @books = Book.limit(params[:limit])
-    render json: @books, meta: { total: Book.count }
-  end
-
-  # GET /books/1
-  def show
-    render json: @book
-  end
-
-  # POST /books
-  def create
-    @book = Book.new(book_params)
-
-    if @book.save
-      render json: @book, status: :created, location: @book
-    else
-      render json: @book.errors, status: :unprocessable_entity
-    end
-  end
-
-  # PATCH/PUT /books/1
-  def update
-    if @book.update(book_params)
-      render json: @book
-    else
-      render json: @book.errors, status: :unprocessable_entity
-    end
-  end
-
-  # DELETE /books/1
-  def destroy
-    @book.destroy
-  end
-
-  private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_book
-      @book = Book.find(params[:id])
-    end
-
-    # Only allow a trusted parameter "white list" through.
-    def book_params
-      res = ActiveModelSerializers::Deserialization.jsonapi_parse(params, polymorphic: [:publisher])
-      res[:publisher_type] = res[:publisher_type].singularize.capitalize
-      res
-    end
-end
