--- bookstore-api/app/models/book.rb
@@ -1,4 +1,0 @@
-class Book < ApplicationRecord
-  belongs_to :author
-  belongs_to :publisher, polymorphic: true
-end
