--- bookstore-api/app/models/publishing_house.rb
@@ -1,3 +1,0 @@
-class PublishingHouse < ApplicationRecord
-  has_many :published, as: :publisher, foreign_key: :publisher_id, class_name: 'Book'
-end
