--- bookstore-api/app/serializers/author_serializer.rb
@@ -1,5 +1,0 @@
-class AuthorSerializer < ActiveModel::Serializer
-  attributes :id, :name, :discount
-  has_many :books
-  has_many :published
-end
