--- bookstore-api/app/serializers/book_serializer.rb
@@ -1,5 +1,0 @@
-class BookSerializer < ActiveModel::Serializer
-  attributes :id, :title, :price
-  belongs_to :author
-  belongs_to :publisher
-end
