--- bookstore-api/app/serializers/publishing_house_serializer.rb
@@ -1,4 +1,0 @@
-class PublishingHouseSerializer < ActiveModel::Serializer
-  attributes :id, :name, :discount
-  has_many :published
-end
